requirements.txt
click==8.1.8
dash-bootstrap-components==1.6.0
dash==2.18.2
glob2==0.7
ipython==8.26.0
kaleido==1.0.0rc0
loguru==0.7.3
matplotlib==3.10.0
medspacy==1.3.1
nbformat==5.10.4
numpy==1.26.4
pandas==2.2.1
plotly-express==0.4.1
poetry==1.2.0
pydantic-settings==2.2.1
pysimstring==1.2.1
python-dotenv==1.0.1
scikit-learn==1.4.2
seaborn==0.13.2
shap==0.45.0
spacy==3.7.5
tensorflow==2.15.0
tqdm==4.67.1
